+++ lib/action_policy/behaviour.rb
@@ -33,11 +33,13 @@ module ActionPolicy
     # Policy is inferred from record
     # (unless explicitly specified through `with` option).
     #
+    # @return the policy record
     # Raises `ActionPolicy::Unauthorized` if check failed.
     def authorize!(record = :__undef__, to:, **options)
       policy = lookup_authorization_policy(record, **options)
 
       Authorizer.call(policy, authorization_rule_for(policy, to))
+      policy.record
     end
 
     # Checks that an activity is allowed for the current context (e.g. user).